Replace if statement
Replaced the if statement checking for the existance of the tile data with an assertion (i.e. the code supposes tile data exists and if it doesnt while the game is still being developed, it crashes to notify the developers), as tile_data should never be *null* as long as the player touches any ground tiles (or even tiles in general). if it is nonetheless, it would mean that the map_position doesn't point to the gorund anymore.