Improve detection code
moved all contents from the previous **check_moving_platform_type** function to the for loop, where it was called directly. this is now possible due to the previous commit decluttering the code such that there is only one function call, which can just be replaced with its contents anyway. this also declutters the code by reducing boilerplate code (i.e. unnecessary code, like in this case, defining a function and calling it, just to do the exact same, specialized task)