Build state machine
Build the state machine scene, for a finite state machine implementation based on nodes. The state machine is a scene with the core code in a script. also created the base state class, which the state machine interacts with. also integrated the state machine into the player scene and called its main functions when necessary