Streamline interactables
Marked the moving platform script and the surprise block script as tool scripts: this is used to reduce the amount of scenes, each with their own small and insignificant variation on the main object, like color or size. Now only streamlined surprise_block and moving_platform scenes exist and the color and/or size can be edited in the inspector as exported variables. Using Godot's tool script functionality, this updates sprites and/or collision shapes inside of the editor environment.