Integrate particles
added a gpu based particle system node. I recommend going into godot with the bridges branch active and to just look at the particles themselves and at the configuration of the node. made it so that the particle system is active only when the bridge is marked as instable. when the color is changed, the base color of the particles also changes accordingly. when the brigde collapses, the particles stop